fixing int vs int64 data type, and fixing cause of incorrect request counter

+++ fdbclient/include/fdbclient/BlobWorkerCommon.h
@@ -47,8 +47,8 @@ struct BlobWorkerStats {
 	Counter forceFlushCleanups;
 	Counter oldFeedSnapshots;
 
-	int numRangesAssigned;
-	int mutationBytesBuffered;
+	int64_t numRangesAssigned;
+	int64_t mutationBytesBuffered;
 	int activeReadRequests;
 	// TODO: add gauge for granules blocking on old snapshots, once this guage is fixed
 	int granulesPendingSplitCheck;
